refactor: :recycle: separation of responsibility according to Solid in insertList route
