+++ src/modules/useCases/book/getListAll/Repositories/IgetListBookAll.Repository.ts
@@ -1,4 +1,4 @@
-import { ListType } from '../../../../types/list';
+import { ListType } from '../../insert/InsertCSVBookUseCase';
 
 export interface IgetListBookAllRepository {
     getListAll(): Promise<ListType[]>
