+++ src/modules/useCases/book/getListAll/Repositories/getListBookAll.Repository.ts
@@ -1,5 +1,5 @@
-import { client } from '../../../../database/client';
-import { ListType } from '../../../../types/list';
+import { client } from '../../../../../database/client';
+import { ListType } from '../../insert/InsertCSVBookUseCase';
 import { IgetListBookAllRepository } from './IgetListBookAll.Repository';
 
 export class GetListBookAllRepository implements IgetListBookAllRepository {
