--- src/types/list.ts
@@ -1,7 +1,0 @@
-export type ListType = {
-    title: string;
-    isbn: string;
-    author: string;
-    category: string;
-    value: number;
-};
